+++ xseof/mpl_orbref/__init__.py
@@ -69,7 +69,7 @@ def load(source):
     """Load a Reference Orbit Event from the source stream.
 
     The input stream can be a filename, a file like object (open in
-    binary mode) or an xml ElementTree.
+    binary mode) or an xml ElementTree (lxml is needed for the latter).
     """
     if hasattr(source, "read"):
         return from_string(source.read())